rf: add lib.rs and move business logic out of the main module
